pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "team-manager"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "aubravo", email = "alvaroubravo@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = ["django", "psycopg2", "django-phonenumber-field[phonenumbers]"]

[project.urls]
Documentation = "https://github.com/aubravo/team-manager#readme"
Issues = "https://github.com/aubravo/team-manager/issues"
Source = "https://github.com/aubravo/team-manager"

[tool.hatch.version]
source = "vcs"

[tool.hatch.envs.default]
type = "container"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/team_manager tests}"

[tool.coverage.run]
source_pkgs = ["team_manager", "tests"]
branch = true
parallel = true
omit = [
  "src/team_manager/__about__.py",
]

[tool.coverage.paths]
team_manager = ["src/team_manager", "*/team-manager/src/team_manager"]
tests = ["tests", "*/team-manager/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
